Update sql definition to autogenerate user ids

+++ db/tables.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS public.user;
 
 CREATE TABLE public.user 
 (
-    id NUMERIC PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name VARCHAR(50) NOT NULL,
     email VARCHAR(50) NOT NULL UNIQUE,
     password VARCHAR(50) NOT NULL
